contract-review-backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyPDF2
python-docx
google-cloud-aiplatform
google-cloud-storage
google-genai
